Trait npm_tools::Visitor

pub trait Visitor {
    type Error;
    fn error(&mut self, package: &PackageInfo, err: &Error);
    fn change(&mut self, action: Instruction) -> Result<(), Self::Error>;
}

Associated Types

type Error

Required Methods

fn error(&mut self, package: &PackageInfo, err: &Error)

Called whenever the package identified by package could be processed. The exact problem is stated in err.

fn change(&mut self, action: Instruction) -> Result<(), Self::Error>

Called with an instruction on what to do next. Must never panic, and is expected to keep all error handling internal.